+++ CHANGELOG.md
@@ -19,6 +19,10 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* Fixed [#238](https://github.com/NeoVintageous/NeoVintageous/issues/238): Simply search & replace not working for me
+* Fixed [#111](https://github.com/NeoVintageous/NeoVintageous/issues/111): Bad command
+* Fixed: "Traling characters" Status message typo
+* Fixed: `:s[ubstitute]` No previous substitute error message is incorrect
 * Fixed: [#226](https://github.com/NeoVintageous/NeoVintageous/issues/226): Mouse does not reset cursor column
 * Fixed: `C-w _` (set current group height as high as possible) doesn't always work correctly
 * Fixed: `C-w |` (set current group width as wide as possible) doesn't always work correctly